Rename to rules
Make loaders rules
