+++ lib/install/template.rb
@@ -1,7 +1,7 @@
 # Install Webpacker
 copy_file "#{__dir__}/config/webpacker.yml", "config/webpacker.yml"
 
-say "Copying webpack core config and loaders"
+puts "Copying webpack core config"
 directory "#{__dir__}/config/webpack", "config/webpack"
 
 say "Copying .postcssrc.yml to app root directory"
