+++ package/index.js
@@ -7,7 +7,7 @@ const merge = require('webpack-merge')
 const Environment = require('./environment')
 const config = require('./config')
 const assetHost = require('./asset_host')
-const loaders = require('./loaders')
+const rules = require('./rules')
 
 const createEnvironment = () => {
   const path = resolve(__dirname, 'environments', `${process.env.NODE_ENV}.js`)
@@ -17,4 +17,4 @@ const createEnvironment = () => {
 
 const environment = createEnvironment()
 
-module.exports = { merge, environment, config, assetHost, loaders, Environment }
+module.exports = { merge, environment, config, assetHost, rules, Environment }
