+++ package/rules/coffee.js
