+++ package/rules/file.js
