+++ package/rules/index.js
