+++ package/rules/typescript.js
